GRN Panel Updates

+++ database/migrations/2025_07_09_100315_create_grn_items_table.php
@@ -6,18 +6,18 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-        public function up(): void
+    public function up(): void
     {
         Schema::create('grn_items', function (Blueprint $table) {
             $table->uuid('id')->primary();
             $table->uuid('grn_id');
             $table->uuid('item_id');
             $table->string('item_name');
-            $table->decimal('qty', 12, 2);
-            $table->decimal('foc', 12, 2)->default(0);
+            $table->unsignedInteger('qty');
+            $table->unsignedInteger('foc')->default(0);
             $table->decimal('price', 12, 2);
-            $table->decimal('margin', 5, 2)->nullable();
-            $table->decimal('discount', 5, 2)->nullable();
+            $table->unsignedInteger('margin')->nullable();
+            $table->unsignedInteger('discount')->nullable();
             $table->decimal('final_price', 12, 2);
             $table->decimal('subtotal', 12, 2);
             $table->timestamps();
